dirty fix for haplotypes merging
