+++ src/projects/trio/run_trio.cpp
@@ -27,9 +27,9 @@ int main(int argc, char **argv) {
     std::experimental::filesystem::path corrected_reads(parser.getValue("corrected_reads"));
 
     io::Library reads_lib = oneline::initialize<std::experimental::filesystem::path>(parser.getListValue("reads"));
-    std::experimental::filesystem::path res_m(dir / "graph_m.gfa");
+    std::experimental::filesystem::path res_m(dir / "graph_p.gfa");
     simplifyHaplo(logger, threads, res_m, graph, haplo, 'm', corrected_reads, reads_lib, dir);
-    std::experimental::filesystem::path res_p(dir / "graph_p.gfa");
+    std::experimental::filesystem::path res_p(dir / "graph_m.gfa");
     simplifyHaplo(logger, threads, res_p, graph, haplo, 'p', corrected_reads, reads_lib, dir);
