+++ src/projects/trio/trio.cpp
@@ -29,43 +29,6 @@ using logging::Logger;
 using std::cout;
 using std::cerr;
 
-inline char other_haplo(char c) {
-    if (c == 'm') return 'p';
-    else if (c == 'p') return 'm';
-    else VERIFY(false);
-}
-
-inline char is_defined_haplo(char c){
-    return (c == 'm' || c == 'p');
-}
-
-struct HaplotypeStats {
-    char haplotype;
-    std::string label;
-//order: p, m, as in file;
-    std::array<int, 2> decisive_strips;
-    std::array<int, 2> decisive_counts;
-    int total_kmers;
-
-//        32         m       0       273     28      390     22      24      112906  17
-//#s->seq[i].name, type, s->cnt[i].sc[0], s->cnt[i].sc[1],c[0<<2|2], c[2<<2|0], c[0<<2|1], c[1<<2|0], s->cnt[i].nk, c[0])
-//TODO: arr 6, 7, 9
-    HaplotypeStats(std::string s) {
-        std::vector<std::string> tokens = ::split(s);
-        haplotype = tokens[1][0];
-        label = tokens[0];
-        decisive_strips = std::array<int,2>{stoi(tokens[2]), stoi(tokens[3])};
-        decisive_counts = std::array<int,2>{stoi(tokens[4]), stoi(tokens[5])};
-        total_kmers = stoi(tokens[8]);
-    }
-    HaplotypeStats():haplotype('u'),label(""), decisive_counts{0,0}, decisive_strips{0,0}, total_kmers(0) {}
-    bool is_undefined() {
-        return (haplotype != 'm' and haplotype != 'p');
-    }
-};
-
-typedef  std::unordered_map<std::string, HaplotypeStats> haplo_map_type;
-//primitive clean graph
 
 void cleanGraph(multigraph::MultiGraph &graph) {
     bool changed = true;
